# api-spec/src/main/resources/static/api.yml
...
swagger: '2.0'
info:
  title: Name of the App API
  description: Use our services using this API
  version: "1.0.0"
# the domain of the service
host: api.app.com
...
schemes:
  - https
# will be prefixed to all paths
basePath: /v1
produces:
  - application/json
paths:
  /users:
    post:
      summary: Create user
      description: |
        This method creates a new user.
      parameters:
        - name: public_identifier
          in: query
          description: Public identifier of the user.
          required: true
          type: string
        - name: email
          in: query
          description: Email of the user.
          required: true
          type: string
        - name: pass
          in: query
          description: Password of the user.
          required: true
          type: string
        - name: first_name
          in: query
          description: First name of the user.
          required: true
          type: string
        - name: last_surname
          in: query
          description: Last name of the user.
          required: true
          type: string
        - name: birthday
          in: query
          description: Date of the user's birthday.
          required: true
          type: string
        - name: imei
          in: query
          description: Unique identifier from the user's device.
          required: true
          type: string
      tags:
        - Users
      responses:
        201:
          description: User created
          schema:
            $ref: '#/definitions/User'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /users/{public_identifier}:
    put:
      summary: Modify user
      description: |
        This method modifies the data of a previously created user.
      parameters:
        - name: public_identifier
          in: path
          description: Public identifier of the user.
          required: true
          type: string
        - name: email
          in: query
          description: Email of the user.
          required: false
          type: string
        - name: pass
          in: query
          description: Password of the user.
          required: false
          type: string
        - name: first_name
          in: query
          description: First name of the user.
          required: false
          type: string
        - name: last_surname
          in: query
          description: Last name of the user.
          required: false
          type: string
        - name: birthday
          in: query
          description: Date of the user's birthday.
          required: false
          type: string
        - name: imei
          in: query
          description: Unique identifier from the user's device.
          required: false
          type: string
      tags:
        - Users
      responses:
        202:
          description: User's data modified
          schema:
            $ref: '#/definitions/User'
        304:
          description: User's data not modified
          schema:
            $ref: '#/definitions/Error'
        404:
          description: User not found
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Delete user
      description: |
        This method deletes the data of a previously created user.
      parameters:
        - name: public_identifier
          in: path
          description: Public identifier of the user.
          required: true
          type: string
      tags:
        - Users
      responses:
        200:
          description: User deleted
          schema:
            $ref: '#/definitions/User'
        404:
          description: User not found
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    get:
      summary: Get user
      description: |
        This method returns a user by public identifier.
      parameters:
        - name: public_identifier
          in: path
          description: Public identifier of the user.
          required: true
          type: string
      tags:
        - Users
      responses:
        200:
          description: User with public identifier requested
          schema:
            $ref: '#/definitions/User'
        404:
          description: User not found
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  User:
    type: object
    properties:
      public_identifier:
        type: string
        description: Public identifier of the user.
      email:
        type: string
        description: Email of the user.
      pass:
        type: string
        description: Password of the user.
      first_name:
        type: string
        description: First name of the user.
      last_surname:
        type: string
        description: Last name of the user.
      birthday:
        type: string
        description: Date of the user's birthday.
      imei:
        type: string
        description: Unique identifier from the user's device.
...
  Error:
    type: object
    properties:
      title:
        type: string
      body:
        type: string